Only download mermaid.js if MERMAID_USE_CDN is False
This is a follow-up to
https://github.com/pysnippet/django-mermaid/issues/20

https://github.com/pysnippet/django-mermaid/pull/21 added a CDN feature,
which means the added script tag will point to the CDN version of
mermaid.js. But it still tries to download a copy to the package's
static directory in AppConfig.ready.

Because this isn't necessary in CDN mode and causes us a PermissionError
during deployment, I'm proposing we only download in non-CDN mode.